Move all style utilities into "styles"
