+++ pages/docs/[component].tsx
@@ -5,7 +5,7 @@ import { Box } from '../../src/components/Box'
 import { Callout, CalloutVariant } from '../../src/components/Callout'
 import { MarkdownWrapperStyles } from '../../src/components/Markdown'
 import { Text, textVariants, TextVariantToken } from '../../src/components/Text'
-import { palette, spacing, SpacingToken } from '../../src/styles/theme'
+import { palette, spacing, SpacingToken } from '../../styles/theme'
 /** @jsx jsx */ jsx
 
 const SectionHeading = ({ children }: { children: ReactNode }) => {
