+++ pages/tags/[tag].tsx
@@ -8,7 +8,7 @@ import { Box } from '../../src/components/Box'
 import { Heading } from '../../src/components/Heading'
 import { Layout } from '../../src/components/Layout'
 import { Text } from '../../src/components/Text'
-import { textMaxWidth } from '../../src/styles/variables'
+import { textMaxWidth } from '../../styles/variables'
 import { compareDesc, format } from 'date-fns'
 import { PostListItem } from '../../src/components/PostListItem'
 import { css, jsx } from '@emotion/core'
