+++ src/components/Section.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled'
-import { textMaxWidth } from '../styles/variables'
-import { spacing } from '../styles/theme'
+import { textMaxWidth } from '../../styles/variables'
+import { spacing } from '../../styles/theme'
 
 export const Section = styled.section({
   margin: `${spacing.m}px auto`,
